Implemented Basic Parry And Health System & Added Comments to Scripts
The Parry does nothing but create a console log that says "Parry!" But it's primed to do anything else once we implement enemies. This is also true for the player health script, it exists in concept, but it doesn't do anything because there's nothing for the player to take damage on. Once we implement things like enemies, these things will work, they are already set up. Also, I added comments hopefully explaining how the scripts work.