Added: Placeholder Character W/ Anims & Melee Attack Code (Semi-functional) & Started on EnemyHealth
I added a placeholder main character with an attack animation so that I could work on the melee attack code. I also added an enemy health script that I will be working on, as well as an animationControllers folder for sorting. The attack code is semi-functional, but doesn't do much of anything right now other than create a debug.log.