Camera Script \ Combat Movement Fixes
Added a functional script for the Main Camera to follow the player in a very nice, pleasing way. It's functional! Try it!! I also made it so that instead of flipping the sprite, the player gameObject itself gets flipped, this is so that the overlap circle can flip along with the player and attack in the direction the player is facing, so now the combat script has even more function. There's also just some updates to my testing scene here.