Movement & Enemy AI Changes
I tweaked the enemy ai and fixed it so that they turn to face the player better (along with their attack hitpoint), and fixed a bug where sometimes they would turn around when the player was approaching them. I also tweaked the movement so jumping hopefully feels less floating and AGAIN messed with how the player is flipped, so instead it's flipping the scale of the player.