Added A Currency System & "Items"
I scripted a currency and item system that allows you to buy items and pick up coins for money.
A few caveats:
The coin doesn't destroy itself on pickup, whenever you walk over it, you get 5 coins, you could do it forever.
The item is a placeholder item I created. You set the item id in the inspector (ItemNumberSet) and that dictates what it does, there's only 2 potential options, so only put one or two.
What the item does: Turns on a bool, destroys itself.
If you are poor, then the debug log will say you are broke!
You can set the players starting amount in the inspector.
You can set the item cost in the inspector.
The coin and the item prefab are in the prefabs folder, have fun! Stay safe, too!
PS: It's totally valid if we want chests that drop random items, let me know if you guys want to do that, I can script it in literally no time.