PlayerAnimation, SwordAnimation, CoinTweaks, Prefab Updates
This ones kind of a doozy, so lock in:
I added player animations to the player prefab, thank you Kevin for creating them.
I also added a sword that he holds and that when you press the attack button (R) it animates downward immediately, kind of clunky but it's temporary. MAKE SURE YOU SET THE SWORD IN THE INSPECTOR FOR THE COMBAT SCRIPT.
I added an item prefab that when you purchase it, it increases how many times you can jump.
MAJOR CHANGE ALERT! Jumping is no longer determined on if you're grounded or not. You have a jumpLimit and JumpMax. Your jumpLimit is equal to your JumpMax when you are grounded, when you are off the ground you can jump as many times as your jumpLimit, when you run out you can't jump anymore until your jumpLimit resets when you are grounded.
This item adds one to your jumpMax, that's all.
I also cleaned up the currency script for items.
I also added a script called CoinBehavior that makes it so that coins disappear when a player picks them up. No more infinite money glitch.
There's also a lot of art in this commit, I tried to sort it, and I think I did a pretty good job, LMK if you have trouble finding something.
There's also some tweaks to the enemy prefab and the enemyAI, as of right now he has no gravity when going off of a ledge, yes I will fix it soon, I'm sorry for introducing it.
There is no parry animation as of this moment! If you are parrying there is no way to tell!