Player User Interface & Coin Drops
I added a Player UI that has a dash cooldown bar and a healthbar that you can set in the inspector. There is a player UI prefab you can drag and drop into your hierarchy. I also made it so that when enemies die they drop coins. I also added a second item that you could set the ITEMNUMBER to; this one raises your attackRate.