Iframes, Health Value Text, Enemy Collisions
You no longer collide with enemies, period, this is so it doesn't mess with the character during gameplay.
Whenever you are hit, you have iframes that prevent you from being hit again, this can be tweaked in the players inspector.
There is now text on the healthbar, showing a number of how much health you have.